added minitest and modified last task
